修复bug
修改视频详情页up主在为获取到数据的默认显示
修改在打开网站时未登录的情况下跳转到登录页
修复在未登录情况下视频详情页up主被后端拦截的问题
修改后端在getUserById中返回UserVO

+++ cilicili-back/src/main/java/com/cilicili/service/UserService.java
@@ -8,7 +8,7 @@ import org.springframework.web.multipart.MultipartFile;
 
 
 public interface UserService  {
-    public User getUserById(int id);
+    public User getUserById(long id);
     public Result<User> login(String username, String password);
     public Result<User> register(String username, String phone, String password);
     String uploadAvatar(MultipartFile file, int userId);
